hf_back/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Google Gemini AI
google-generativeai==0.3.2

# Document Generation
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.1.0

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1
weasyprint==60.1

# HTML to PDF
pdfkit==1.0.0

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# JSON Processing
jsonschema==4.20.0

# Utilities
python-magic==0.4.27
